refactor(Velocity): move to the movement category
